refactor(slack): split normalizers by event family; delete normalize.ts
Final structural step of the slack/normalize.ts reorg. Split the normalizer
functions out of the last monolith into one module per event family and
remove normalize.ts:

- message-normalizer.ts        — app_mention / DM / channel + the shared
                                  buildNormalizedSlackMessage builder
- message-change-normalizer.ts — message edit + delete
- reaction-normalizer.ts       — reaction_added / reaction_removed
- block-actions-normalizer.ts  — block_actions interactive payloads

normalize.ts is renamed to message-normalizer.ts (git mv) with the other
three families cut out, so history follows the bulk. Consumers (socket-mode
and the normalizer tests) import each function from its family module.

Pure move — no logic change. 298 slack tests pass; tsc / eslint / knip
clean. With user-directory (#39028), the shared helpers (#39037), and these
family modules, slack/ now matches the Module Organization convention and
normalize.ts is gone.
